Moved v1_ops to their own  folder that is omitted from coverage. Linear activation is it's own layer. Corrected weighted shuffle algorithm to use inverse probability
