+++ beagles/backend/net/ops/__init__.py
@@ -1,14 +1,28 @@
 from beagles.backend.net.ops.simple import *
+from beagles.backend.net.ops.rnn import *
 from beagles.backend.net.ops.convolution import *
 
 op_types = {
     'convolutional': Convolutional,
     'connected': Connected,
+    'conv-lstm': ConvLSTM,
+    'lstm': LSTM,
+    'rnn': RNN,
+    'gru': GRU,
     'maxpool': MaxPool,
     'stair': Stair,
     'hardtan': HardTan,
     'relu': Relu,
+    'relie': Relie,
+    'linear': Linear,
+    'logistic': Logistic,
+    'loggy': Loggy,
+    'gelu': Gelu,
+    'selu': Selu,
     'elu': Elu,
+    'psle': PSLE,
+    'tanh': Tanh,
+    'lhtan': LHTan,
     'leaky': Leaky,
     'shortcut': Shortcut,
     'upsample': UpSample,
